Tests for blueprint_meta_controller
