+++ src/opera/api/tests/test_09_blueprint.py
@@ -288,156 +288,3 @@ class TestUser:
         assert_that(resp.status_code).is_equal_to(200)
         resp = client.get(f"/blueprint/{uuid.uuid4()}/user")
         assert_that(resp.json).is_empty()
-
-
-class TestValidateExisting:
-
-    def test_no_blueprint(self, client, mocker):
-        blueprint_id = uuid.uuid4()
-        mock_version_exists = mocker.MagicMock(name='version_exists', return_value=False)
-        mocker.patch('opera.api.service.csardb_service.GitDB.version_exists', new=mock_version_exists)
-
-        resp = client.put(f"blueprint/{blueprint_id}/validate")
-        assert resp.status_code == 404
-        mock_version_exists.assert_called_with(str(blueprint_id), None)
-
-    def test_exception(self, client, mocker):
-        blueprint_id = uuid.uuid4()
-        mocker.patch('opera.api.service.csardb_service.GitDB.version_exists', return_value=True)
-        mock_validate = mocker.MagicMock(name='validate',
-                                         return_value="{}: {}".format(Exception.__class__.__name__,
-                                                                      "Exception stacktrace"))
-
-        mocker.patch('opera.api.controllers.background_invocation.InvocationWorkerProcess.validate', new=mock_validate)
-        resp = client.put(f"/blueprint/{blueprint_id}/validate")
-
-        assert resp.status_code == 200
-        assert_that(resp.json).contains_only("blueprint_valid", "error")
-        assert_that(resp.json['blueprint_valid']).is_false()
-        assert_that(resp.json['error']).contains("Exception stacktrace")
-        mock_validate.assert_called_with(str(blueprint_id), None, None)
-
-    def test_ok(self, client, mocker, inputs_1):
-        blueprint_token = uuid.uuid4()
-        mocker.patch('opera.api.service.csardb_service.GitDB.version_exists', return_value=True)
-        mock_validate = mocker.MagicMock(name='validate', return_value=None)
-        mocker.patch('opera.api.controllers.background_invocation.InvocationWorkerProcess.validate', new=mock_validate)
-        resp = client.put(f"/blueprint/{blueprint_token}/validate", data=inputs_1)
-
-        assert resp.status_code == 200
-        assert_that(resp.json).contains_only("blueprint_valid")
-        assert_that(resp.json['blueprint_valid']).is_true()
-        mock_validate.assert_called_with(str(blueprint_token), None, {'marker': 'blah'})
-
-
-class TestValidateExistingVersion:
-
-    def test_no_blueprint(self, client, mocker):
-        blueprint_id = uuid.uuid4()
-        version_id = 'v1.0'
-        mock_version_exists = mocker.MagicMock(name='version_exists', return_value=False)
-        mocker.patch('opera.api.service.csardb_service.GitDB.version_exists', new=mock_version_exists)
-
-        resp = client.put(f"blueprint/{blueprint_id}/version/{version_id}/validate")
-        assert resp.status_code == 404
-        mock_version_exists.assert_called_with(str(blueprint_id), version_id)
-
-    def test_exception(self, client, mocker):
-        blueprint_id = uuid.uuid4()
-        version_id = 'v1.0'
-        mocker.patch('opera.api.service.csardb_service.GitDB.version_exists', return_value=True)
-        mock_validate = mocker.MagicMock(name='validate',
-                                         return_value="{}: {}".format(Exception.__class__.__name__,
-                                                                      "Exception stacktrace"))
-        mocker.patch('opera.api.controllers.background_invocation.InvocationWorkerProcess.validate', new=mock_validate)
-        resp = client.put(f"/blueprint/{blueprint_id}/version/{version_id}/validate")
-
-        assert resp.status_code == 200
-        assert_that(resp.json).contains_only("blueprint_valid", "error")
-        assert_that(resp.json['blueprint_valid']).is_false()
-        assert_that(resp.json['error']).contains("Exception stacktrace")
-        mock_validate.assert_called_with(str(blueprint_id), version_id, None)
-
-    def test_ok(self, client, mocker, inputs_1):
-        blueprint_token = uuid.uuid4()
-        version_id = 'v1.0'
-        mocker.patch('opera.api.service.csardb_service.GitDB.version_exists', return_value=True)
-        mock_validate = mocker.MagicMock(name='validate', return_value=None)
-        mocker.patch('opera.api.controllers.background_invocation.InvocationWorkerProcess.validate', new=mock_validate)
-        resp = client.put(f"/blueprint/{blueprint_token}/version/{version_id}/validate", data=inputs_1)
-
-        assert resp.status_code == 200
-        assert_that(resp.json).contains_only("blueprint_valid")
-        assert_that(resp.json['blueprint_valid']).is_true()
-        mock_validate.assert_called_with(str(blueprint_token), version_id, {'marker': 'blah'})
-
-
-class TestValidateNew:
-
-    def test_exception(self, client, mocker, csar_1):
-        mocker.patch('opera.api.controllers.background_invocation.InvocationWorkerProcess.validate_new',
-                     return_value="{}: {}".format(Exception.__class__.__name__, "Exception stacktrace"))
-        resp = client.put(f"/blueprint/validate", data=csar_1)
-
-        assert resp.status_code == 200
-        assert_that(resp.json).contains_only("blueprint_valid", "error")
-        assert_that(resp.json['blueprint_valid']).is_false()
-        assert_that(resp.json['error']).contains("Exception stacktrace")
-
-    def test_ok(self, client, mocker, csar_1):
-        mocker.patch('opera.api.controllers.background_invocation.InvocationWorkerProcess.validate', return_value=None)
-        resp = client.put(f"/blueprint/validate", data=csar_1)
-
-        assert resp.status_code == 200
-        assert_that(resp.json).contains_only("blueprint_valid")
-        assert_that(resp.json['blueprint_valid']).is_true()
-
-
-class TestGitHistory:
-
-    def test_not_found(self, client, mocker, patch_auth_wrapper):
-        mocker.patch('opera.api.service.sqldb_service.Database.get_git_transaction_data', return_value=None)
-        blueprint_id = uuid.uuid4()
-        resp = client.get(f"/blueprint/{blueprint_id}/git_history")
-        assert resp.status_code == 400
-        assert_that(resp.json).contains('Log not found')
-
-    def test_no_blueprint(self, client, mocker):
-        # assert endpoint works even when blueprint is gone
-        blueprint_id = uuid.uuid4()
-        mocker.patch('opera.api.service.csardb_service.GitDB.version_exists', return_value=False)
-        git_data = GitLog(
-            blueprint_id=str(blueprint_id),
-            commit_sha="commit_sha",
-            git_backend="MockConnector",
-            job="update",
-            repo_url="local",
-            revision_msg="rev_msg",
-            timestamp=timestamp_util.datetime_now_to_string(),
-            version_id='v1.0'
-        )
-        mocker.patch('opera.api.service.sqldb_service.Database.get_git_transaction_data',
-                     return_value=[git_data.to_dict()])
-        blueprint_id = uuid.uuid4()
-        resp = client.get(f"/blueprint/{blueprint_id}/git_history")
-        assert resp.status_code == 200
-
-    def test_keys(self, client, mocker, patch_auth_wrapper):
-        git_data = GitLog(
-            blueprint_id=str(uuid.uuid4()),
-            commit_sha="commit_sha",
-            git_backend="MockConnector",
-            job="update",
-            repo_url="local",
-            revision_msg="rev_msg",
-            timestamp=timestamp_util.datetime_now_to_string(),
-            version_id='v1.0'
-        )
-        mock_git_data = mocker.MagicMock(name='invoke', return_value=[git_data.to_dict()])
-        mocker.patch('opera.api.service.sqldb_service.Database.get_git_transaction_data', new=mock_git_data)
-
-        resp = client.get(f"/blueprint/{git_data.blueprint_id}/git_history")
-        assert resp.status_code == 200
-        assert_that(resp.json).is_length(1)
-        assert_that(resp.json[0]).contains_only(*git_data.to_dict().keys())
-        mock_git_data.assert_called_with(git_data.blueprint_id, fetch_all=True)
